Test refactoring & AOC2019 day 9 part 1 & 2 completed!

+++ include/intcode_computer.hpp
@@ -6,7 +6,7 @@
 #include <queue>
 #include <functional>
 
-using Code = long;
+using Code = long long;
 
 enum class Opcode {
     unknown = -1,
